Cleaned up the code and added comments for the controllers' endpoints.

+++ src/main/java/com/COMP3095/gbc_pay/controllers/UserDashboardController.java
@@ -14,34 +14,35 @@ import com.COMP3095.gbc_pay.services.dashboard.MessageService;
 import com.COMP3095.gbc_pay.services.dashboard.user.CreditProfileService;
 import com.COMP3095.gbc_pay.services.dashboard.user.UserProfileService;
 import com.COMP3095.gbc_pay.validation.CustomValidationService;
-import org.springframework.context.support.DefaultMessageSourceResolvable;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collectors;
+
 
 
 @Controller
 @RequestMapping({"/dashboard", "/dashboard.html"})
-@SessionAttributes({"currentProfile", "currentViewProfile", "allProfiles"})
+@SessionAttributes({"currentProfile", "currentViewProfile", "allProfiles", "portNumber"})
 public class UserDashboardController {
 
     private final UserProfileService userProfileService;
     private final CreditProfileService creditProfileService;
     private final MessageService messageService;
     private final CustomValidationService customValidationService;
+
+    @Value("${server.port}")
+    private int portNumber;
 
     public UserDashboardController(UserProfileService userProfileService, CreditProfileService creditProfileService,
                                    MessageService messageService, CustomValidationService customValidationService) {
@@ -51,10 +52,20 @@ public class UserDashboardController {
         this.customValidationService = customValidationService;
     }
 
+    /**
+     *
+     * GET endpoint for the dashboard index page for the USER role.
+     *
+     * @return
+     *  Dashboard Index page (home page).
+     *
+     */
     @GetMapping({"", "/index.html", "index"})
     public String getIndexPage(Model model){
 
         model.addAttribute("currentProfile", userProfileService.getAuthenticatedProfile());
+
+        model.addAttribute("portNumber", portNumber);
 
         model.addAttribute("lastLogin",
                 userProfileService.getLastLogin().format(DateTimeFormatter.ofPattern("EEEE, LLLL dd, yyyy")));
@@ -72,8 +83,15 @@ public class UserDashboardController {
         return "dashboard/user/index";
     }
 
+    /**
+     *
+     * GET endpoint for the My Profile page.
+     *
+     * @return
+     *  My Profile page.
+     *
+     */
     @GetMapping({"my_profile", "my_profile.html"})
-    @PreAuthorize("hasRole('ROLE_USER')")
     public String getMyProfilePage(HttpServletRequest request, Model model, @ModelAttribute("updateEmail") String updateEmail,
                                    @ModelAttribute("user") User user, @ModelAttribute("profile") Profile profile,
                                    @ModelAttribute("address") Address address){
@@ -90,22 +108,26 @@ public class UserDashboardController {
         model.addAttribute("profile", authProfile);
         model.addAttribute("address", authProfile.getAddress());
 
-        System.out.println("The view email is " + email);
-
-
         model.addAttribute("currentViewProfile", authProfile);
         model.addAttribute("allProfiles", userProfileService.getAllProfiles());
 
         return "dashboard/user/my_profile";
     }
 
+    /**
+     *
+     * POST endpoint for My Profile page to delete users.
+     *
+     * @return
+     *  My Profile page on success and failure.
+     *
+     */
     @PostMapping({"my_profile/delete", "my_profile.html/delete"})
     @PreAuthorize("hasRole('ROLE_USER')")
     public String deleteProfile(HttpServletRequest request, RedirectAttributes ra, Model model){
 
         String deleteEmail = request.getParameter("delete-email");
         Profile profile = userProfileService.getProfileByEmail(deleteEmail);
-        System.out.println("The delete email is " + deleteEmail);
 
         if(profile.getEmail().equals(userProfileService.getAuthenticatedProfile().getEmail())){
             ra.addFlashAttribute("error", "You cannot delete the current logged in profile.");
@@ -124,6 +146,14 @@ public class UserDashboardController {
         return "redirect:/dashboard/my_profile";
     }
 
+    /**
+     *
+     * POST endpoint for My Profile to update a profile.
+     *
+     * @return
+     *  My Profile page on success and failure.
+     *
+     */
     @PostMapping({"my_profile", "my_profile.html"})
     @PreAuthorize("hasRole('ROLE_USER')")
     public String updateProfile(RedirectAttributes ra, @ModelAttribute("user") @Valid User user,
@@ -132,7 +162,6 @@ public class UserDashboardController {
 
 
         String updateEmail = ((Profile) Objects.requireNonNull(model.getAttribute("currentViewProfile"))).getEmail();
-
 
         List<String> formErrors = customValidationService.removeErrors(userResult, profileResult, addressResult);
 
@@ -156,7 +185,14 @@ public class UserDashboardController {
         return "redirect:/dashboard/my_profile";
     }
 
-
+    /**
+     *
+     * GET endpoint for the Credit Profile page.
+     *
+     * @return
+     *  Credit Profile page.
+     *
+     */
     @GetMapping({"credit_profile", "credit_profile.html"})
     @PreAuthorize("hasRole('ROLE_USER')")
     public String getCreditProfilePage(HttpServletRequest request, Model model,
@@ -165,7 +201,6 @@ public class UserDashboardController {
         String cardIdNumber = request.getParameter("cardIdNumber");
 
         if(cardIdNumber != null){
-            System.out.println("Card Number is " + cardIdNumber);
 
             Profile tempProfile = (Profile)(model.getAttribute("currentProfile"));
 
@@ -178,6 +213,14 @@ public class UserDashboardController {
         return "dashboard/user/credit_profile";
     }
 
+    /**
+     *
+     * POST end point for the Credit Profile page, to add or update a credit profile.
+     *
+     * @return
+     *  Credit profile page on success or failure.
+     *
+     */
     @PostMapping({"credit_profile", "credit_profile.html"})
     @PreAuthorize("hasRole('ROLE_USER')")
     public String addOrUpdateCredit(Model model, RedirectAttributes ra, @ModelAttribute("card") @Valid Credit credit,
@@ -188,26 +231,36 @@ public class UserDashboardController {
             return "dashboard/user/credit_profile";
         }
 
-        System.out.println("Adding or updating credit for " + credit.getCardHolderNumber());
-
         Profile currProfile = (Profile)(model.getAttribute("currentProfile"));
         creditProfileService.AddOrUpdateCard(currProfile, credit);
-
 
         model.addAttribute("currentProfile", userProfileService.getAuthenticatedProfile());
         ra.addFlashAttribute("updatedCreditProfile", "Credit Profile has been updated.");
         return "redirect:/dashboard/credit_profile";
     }
 
+    /**
+     *
+     * POST endpoint to delete credit profile.
+     *
+     * @return
+     *  Redirect to Credit Profile page on success or failure.
+     *
+     */
     @PostMapping({"credit_profile/delete", "credit_profile.html/delete"})
     @PreAuthorize("hasRole('ROLE_USER')")
     public String deleteCredit(HttpServletRequest request, RedirectAttributes ra, Model model){
 
         String deleteCreditId = request.getParameter("delete-id");
-        System.out.println("The delete id is " + deleteCreditId);
 
         Profile currProfile = (Profile)(model.getAttribute("currentProfile"));
         Credit deleteCard = creditProfileService.getCreditById(currProfile, deleteCreditId);
+
+        if(deleteCard == null){
+            ra.addFlashAttribute("error", "Credit Card could not be deleted.");
+            return "redirect:/dashboard/credit_profile";
+        }
+
 
         creditProfileService.deleteCard(currProfile, deleteCard);
 
@@ -216,6 +269,14 @@ public class UserDashboardController {
         return "redirect:/dashboard/credit_profile";
     }
 
+    /**
+     *
+     * GET endpoint for the Inbox page, support messages that have been sent and the replies.
+     *
+     * @return
+     *  Inbox Page.
+     *
+     */
     @GetMapping({"inbox", "inbox.html"})
     @PreAuthorize("hasRole('ROLE_USER')")
     public String getInboxPage(Model model){
@@ -225,28 +286,39 @@ public class UserDashboardController {
         return "dashboard/user/inbox";
     }
 
+    /**
+     *
+     * GET endpoint to update the read status of the message.
+     *
+     */
     @GetMapping({"inbox/read/{messageId}", "inbox.html/read/{messageId}"})
     @PreAuthorize("hasRole('ROLE_USER')")
     @ResponseStatus(value = HttpStatus.OK)
     public void getReadMessage(@PathVariable(name="messageId", required = false) Integer messageId, Model model){
-        System.out.println("The current read message is " + messageId);
 
         messageService.readMessage(messageId);
 
         model.addAttribute("currentProfile", userProfileService.getAuthenticatedProfile());
     }
 
+    /**
+     *
+     * GET endpoint to delete a message by message id.
+     *
+     * @return
+     *  Inbox Page on success or faliure.
+     *
+     */
     @GetMapping({"inbox/delete/{messageId}", "inbox.html/delete/{messageId}"})
     @PreAuthorize("hasRole('ROLE_USER')")
     public String deleteMessage(@PathVariable(name="messageId", required = false) Integer messageId,
                                 RedirectAttributes ra, Model model){
 
-        System.out.println("Deleting Message with the id: " + messageId);
-
         Profile currProfile = (Profile)(model.getAttribute("currentProfile"));
 
         if(messageId == null || messageService.getMessageById(currProfile, messageId) == null){
             ra.addFlashAttribute("error", "Message could not be deleted.");
+            return "redirect:/dashboard/inbox";
         }
 
         messageService.deleteMessage(currProfile, messageId);
@@ -256,16 +328,31 @@ public class UserDashboardController {
         return "redirect:/dashboard/inbox";
     }
 
+    /**
+     *
+     * GET endpoint for the Support Page to send support messages.
+     *
+     * @return
+     *  Support Page
+     */
     @GetMapping({"support", "support.html"})
     @PreAuthorize("hasRole('ROLE_USER')")
     public String getSupportPage(@ModelAttribute("message") Message message){
         return "dashboard/user/support";
     }
 
+    /**
+     *
+     * POST endpoint to send message with validation.
+     *
+     * @return
+     *  Support page on failure.
+     *  Inbox Page on success.
+     *
+     */
     @PostMapping({"support", "support.html"})
     @PreAuthorize("hasRole('ROLE_USER')")
     public String sendMessage(Model model, RedirectAttributes ra, @ModelAttribute("message") @Valid Message message, BindingResult result){
-        System.out.println("Sending a message with subject: " + message.getSubject());
 
          if(result.hasFieldErrors("subject") || result.hasFieldErrors("messageBody")){
              return "dashboard/user/support";
@@ -278,33 +365,6 @@ public class UserDashboardController {
         ra.addFlashAttribute("messageSent", "Your message has been sent.");
         return "redirect:/dashboard/inbox";
     }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 }
